Handle stupid astroplan bug
