+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "wintertoo"
-version = "1.6.2"
+version = "1.7.0"
 description = ""
 authors = [
     {name = "Robert Stein", email = "rdstein@caltech.edu"},
